requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
fastapi
uvicorn[standard]
python-multipart
pillow
